+++ openapi-operation-validator/src/main/java/org/openapi4j/operation/validator/model/Response.java
@@ -3,11 +3,12 @@ package org.openapi4j.operation.validator.model;
 import org.openapi4j.operation.validator.model.impl.Body;
 
 import java.util.Collection;
+import java.util.Map;
 import java.util.Optional;
 
 public interface Response {
   /**
-   * @return The response status code
+   * @return The response status code.
    */
   int getStatus();
 
@@ -17,13 +18,19 @@ public interface Response {
   Body getBody();
 
   /**
+   * Get the headers.
+   *
+   * @return The headers.
+   */
+  Map<String, Collection<String>> getHeaders();
+
+  /**
    * Get the collection of header values for the header param with the given name.
    *
    * @param name The (case insensitive) name of the parameter to retrieve
    * @return The header values for that param; or empty list. Must be {@code nonnull}.
    */
-  Collection<String> getHeaderValues(String name);
-
+  Collection<String> getHeaderValues(final String name);
 
   /**
    * Get the first of header value for the header param with the given name (if any exist).
@@ -32,11 +39,15 @@ public interface Response {
    * @return The first header value for that param (if it exists)
    */
   default Optional<String> getHeaderValue(final String name) {
-    return getHeaderValues(name).stream().findFirst();
+    Collection<String> values = getHeaderValues(name);
+    if (values != null) {
+      return values.stream().findFirst();
+    }
+    return Optional.empty();
   }
 
   /**
-   * Get the content-type header of this request, if it has been set.
+   * Get the content-type header of this response, if it has been set.
    *
    * @return The content-type header, or empty if it has not been set.
    */
