--- openapi-operation-validator/src/main/java/org/openapi4j/operation/validator/validation/ResponseValidator.java
@@ -1,8 +1,0 @@
-package org.openapi4j.operation.validator.validation;
-
-/**
- * Validate a response data against a given OpenAPI Operation defined in the Spec.
- * Be sure to re-use it each time you need validation since the Operation validators are cached.
- */
-public class ResponseValidator {
-}
